refactor: 메서드 및 클래스 분리
