+++ src/test/java/subway/LineAcceptanceTest.java
@@ -1,19 +1,16 @@
 package subway;
 
 
-import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
 
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 
+import static helper.LineTestHelper.*;
 import static helper.StationTestHelper.지하철역을_생성한다;
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -31,23 +28,13 @@ class LineAcceptanceTest {
         지하철역을_생성한다("지하철역");
         지하철역을_생성한다("새로운지하철역");
 
-        Map<String, Object> params = new HashMap<>();
-        params.put("name", "신분당선");
-        params.put("color", "bg-red-600");
-        params.put("upStationId", 1);
-        params.put("downStationId", 2);
-        params.put("distance", 10);
-        ExtractableResponse<Response> response = 지하철노선을_생성한다(params);
-
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.CREATED.value());
-
-        String lineId = response.jsonPath()
+        String lineId = 지하철노선을_생성한다("신분당선", "bg-red-600", 1L, 2L, 10L)
+                .jsonPath()
                 .get("id")
                 .toString();
 
         // then
         List<String> lines = 지하철노선_목록을_조회한다();
-
         assertThat(lines).contains(lineId);
     }
 
@@ -64,21 +51,8 @@ class LineAcceptanceTest {
         지하철역을_생성한다("새로운지하철역");
         지하철역을_생성한다("또다른지하철역");
 
-        Map<String, Object> line1 = new HashMap<>();
-        line1.put("name", "신분당선");
-        line1.put("color", "bg-red-600");
-        line1.put("upStationId", 1);
-        line1.put("downStationId", 2);
-        line1.put("distance", 10);
-        지하철노선을_생성한다(line1);
-
-        Map<String, Object> line2 = new HashMap<>();
-        line2.put("name", "분당선");
-        line2.put("color", "bg-red-600");
-        line2.put("upStationId", 1);
-        line2.put("downStationId", 3);
-        line2.put("distance", 10);
-        지하철노선을_생성한다(line2);
+        지하철노선을_생성한다("신분당선", "bg-red-600", 1L, 2L, 10L);
+        지하철노선을_생성한다("분당선", "bg-red-600", 1L, 3L, 10L);
 
         // when
         List<String> lines = 지하철노선_목록을_조회한다();
@@ -99,14 +73,7 @@ class LineAcceptanceTest {
         지하철역을_생성한다("지하철역");
         지하철역을_생성한다("새로운지하철역");
 
-        Map<String, Object> line = new HashMap<>();
-        line.put("name", "신분당선");
-        line.put("color", "bg-red-600");
-        line.put("upStationId", 1);
-        line.put("downStationId", 2);
-        line.put("distance", 10);
-
-        String lineId = 지하철노선을_생성한다(line)
+        String lineId = 지하철노선을_생성한다("신분당선", "bg-red-600", 1L, 2L, 10L)
                 .jsonPath()
                 .get("id")
                 .toString();
@@ -130,23 +97,13 @@ class LineAcceptanceTest {
         지하철역을_생성한다("지하철역");
         지하철역을_생성한다("새로운지하철역");
 
-        Map<String, Object> line = new HashMap<>();
-        line.put("name", "신분당선");
-        line.put("color", "bg-red-600");
-        line.put("upStationId", 1);
-        line.put("downStationId", 2);
-        line.put("distance", 10);
-
-        String lineId = 지하철노선을_생성한다(line)
+        String lineId = 지하철노선을_생성한다("신분당선", "bg-red-600", 1L, 2L, 10L)
                 .jsonPath()
                 .get("id")
                 .toString();
 
         // when
-        Map<String, Object> updateLine = new HashMap<>();
-        updateLine.put("name", "신분당선");
-        updateLine.put("color", "bg-red-600");
-        ExtractableResponse<Response> response = 지하철노선을_수정한다(lineId, updateLine);
+        ExtractableResponse<Response> response = 지하철노선을_수정한다(lineId, "신분당선", "bg-red-600");
 
         // then
         assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
@@ -164,14 +121,7 @@ class LineAcceptanceTest {
         지하철역을_생성한다("지하철역");
         지하철역을_생성한다("새로운지하철역");
 
-        Map<String, Object> line = new HashMap<>();
-        line.put("name", "신분당선");
-        line.put("color", "bg-red-600");
-        line.put("upStationId", 1);
-        line.put("downStationId", 2);
-        line.put("distance", 10);
-
-        String lineId = 지하철노선을_생성한다(line)
+        String lineId = 지하철노선을_생성한다("신분당선", "bg-red-600", 1L, 2L, 10L)
                 .jsonPath()
                 .get("id")
                 .toString();
@@ -181,59 +131,5 @@ class LineAcceptanceTest {
 
         // then
         assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
-    }
-
-    private static ExtractableResponse<Response> 지하철노선을_생성한다(Map<String, Object> params) {
-        return RestAssured.given()
-                .body(params)
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .when()
-                .post("/lines")
-                .then()
-                .log().all()
-                .extract();
-    }
-
-    private static List<String> 지하철노선_목록을_조회한다() {
-        return RestAssured.given()
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .when()
-                .get("/lines")
-                .then()
-                .log().all()
-                .extract()
-                .jsonPath()
-                .getList("id", String.class);
-    }
-
-    private ExtractableResponse<Response> 지하철노선을_조회한다(String lineId) {
-        return RestAssured.given()
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .when()
-                .get("/lines/" + lineId)
-                .then()
-                .log().all()
-                .extract();
-    }
-
-    private ExtractableResponse<Response> 지하철노선을_수정한다(String lineId, Map<String, Object> params) {
-        return RestAssured.given()
-                .body(params)
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .when()
-                .post("/lines/" + lineId)
-                .then()
-                .log().all()
-                .extract();
-    }
-
-    public static ExtractableResponse<Response> 지하철노선을_삭제한다(String stationId) {
-        return RestAssured.given()
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .when()
-                .delete("/lines/" + stationId)
-                .then()
-                .log().all()
-                .extract();
     }
 }
